v 0.5

+++ src/main/java/jaime/modelos/MyIDStore.java
@@ -5,7 +5,7 @@ import java.util.concurrent.locks.ReentrantLock;
 
 public class MyIDStore {
     private static MyIDStore instance=null;
-    private int cont=1;
+    private Long cont= Long.valueOf(1);
     private final Lock lockCont = new ReentrantLock(true);
     private MyIDStore(){
 
@@ -17,7 +17,7 @@ public class MyIDStore {
         return instance;
     }
 
-    public int addandgetID(){
+    public Long addandgetID(){
         lockCont.lock();
         try {
             return this.cont++;
